changed quantities from int to long
- Changed transactions variables that hold quantity from int to long
- Same for assets

+++ gerenciador-investimentos/src/main/kotlin/com/vitorbasso/gerenciadorinvestimentos/domain/concrete/Transaction.kt
@@ -23,7 +23,7 @@ data class Transaction(
 
     val type: TransactionType = TransactionType.BUY,
 
-    val quantity: Int = 0,
+    val quantity: Long = 0,
 
     val value: BigDecimal = BigDecimal.ZERO,
 
@@ -31,9 +31,9 @@ data class Transaction(
 
     val checkingValue: BigDecimal = BigDecimal.ZERO,
 
-    val checkingQuantity: Int = 0,
+    val checkingQuantity: Long = 0,
 
-    val daytradeQuantity: Int = 0,
+    val daytradeQuantity: Long = 0,
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "asset_id", referencedColumnName = "id")
